Restructured UTs of deployment-hooks
