+++ broker/applications/deployment_hooks/test/mocha_deployment_hooks.opts
@@ -1,0 +1,6 @@
+--require applications/deployment_hooks/test/support
+--reporter spec
+--check-leaks
+--slow 100
+--growl
+-S
