--- broker/test/test_deployment_hooks/mocha_deployment_hooks.opts
@@ -1,6 +1,0 @@
---require test/test_deployment_hooks/support
---reporter spec
---check-leaks
---slow 100
---growl
--S
